Update SpotBugs plugin and configuration in build-quality.gradle
Replaced the old SpotBugs plugin setup with a new one, and updated the plugin's version to '4.2.3' in build-quality.gradle file. The SpotBugs task configuration has also been modified to enhance the plugin's capabilities and improve the code quality checks. The configuration changes include maintaining the report at a 'high' level while putting in a 'max' effort.